pyproject.toml
[project]
name = "cyber-risk-perception"
version = "0.1.0"
description = ""
authors = [
    {name = "Jean Bertholat",email = "jean.bertholat@alten.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "google-oauth2-tool (>=0.0.3,<0.0.4)",
    "google-api-python-client (>=2.168.0,<3.0.0)",
    "pyvis (>=0.3.2,<0.4.0)",
    "networkx (>=3.4.2,<4.0.0)"
]

packages = [{ include = "cyber_risk_perception", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
